# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Image Processor API",
    "description": "API for image upload, transformation, and management",
    "version": "1.0.0"
  },
  "host": "localhost:5000",
  "basePath": "/api",
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. Example: 'Bearer {token}'"
    },
    "ApiKey": {
      "type": "apiKey",
      "name": "X-API-Key",
      "in": "header",
      "description": "API Key for authentication. Example: 'sk_live_abc123...'"
    }
  },
  "paths": {
...
    "/images/{image_id}/transform": {
      "get": {
        "tags": ["Images"],
        "summary": "Transform an image",
        "security": [{ "Bearer": [] }, { "ApiKey": [] }],
        "parameters": [
          { "in": "path", "name": "image_id", "type": "integer", "required": true },
          { "in": "query", "name": "width", "type": "integer" },
          { "in": "query", "name": "height", "type": "integer" },
          { "in": "query", "name": "format", "type": "string", "enum": ["jpeg", "png", "webp"] },
          { "in": "query", "name": "quality", "type": "integer", "default": 85 },
          { "in": "query", "name": "crop_x", "type": "integer" },
          { "in": "query", "name": "crop_y", "type": "integer" },
          { "in": "query", "name": "crop_width", "type": "integer" },
          { "in": "query", "name": "crop_height", "type": "integer" },
          { "in": "query", "name": "rotate", "type": "integer", "enum": [90, 180, 270] },
          { "in": "query", "name": "watermark", "type": "string" },
          { "in": "query", "name": "optimize", "type": "boolean" },
          { "in": "query", "name": "enhance", "type": "boolean" },
          { "in": "query", "name": "compress", "type": "boolean" },
          { "in": "query", "name": "grayscale", "type": "boolean" },
          { "in": "query", "name": "download", "type": "boolean", "default": false }
        ],
        "responses": {
          "200": { "description": "Transformed image" },
          "404": { "description": "Image not found" }
        }
      }
    },
...
    "/batch/transform": {
      "post": {
        "tags": ["Batch Processing"],
        "summary": "Apply same transformation to multiple images",
        "security": [{"Bearer": []}, {"ApiKey": []}],
        "parameters": [
          {"in": "body", "name": "body", "required": true, "schema": {
            "type": "object",
            "required": ["image_ids"],
            "properties": {
              "image_ids": {"type": "array", "items": {"type": "integer"}, "description": "Array of image IDs (max 10)"},
              "width": {"type": "integer"},
              "height": {"type": "integer"},
              "format": {"type": "string", "enum": ["jpeg", "png", "webp"]},
              "quality": {"type": "integer", "minimum": 1, "maximum": 100},
              "rotate": {"type": "integer", "enum": [90, 180, 270]},
              "watermark": {"type": "string"},
              "optimize": {"type": "boolean"},
              "enhance": {"type": "boolean"},
              "compress": {"type": "boolean"},
              "grayscale": {"type": "boolean"}
            }
          }}
        ],
        "responses": {
          "200": {"description": "ZIP file with transformed images"},
          "400": {"description": "Invalid request"}
        }
      }
    },
...
  }
}